TIP-860: move abstract mass edit processor
