+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/MassEdit/AddProductValueProcessor.php
@@ -3,7 +3,6 @@
 namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit;
 
 use Akeneo\Tool\Component\StorageUtils\Updater\PropertyAdderInterface;
-use Pim\Bundle\EnrichBundle\Connector\Processor\AbstractProcessor;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
